seperated tests for invoice details, changed Fact to Theory for some, added test for checking table

+++ ToyBlockFactoryKata/InvoiceLine.cs
@@ -1,8 +1,8 @@
 namespace ToyBlockFactoryKata
 {
-    public record InvoiceLine(string Description, int Quantity, int Price)
+    public record InvoiceLine(string Description, int  Quantity, int Price)
     {
-        public int Total => Quantity * Price;
+        public decimal Total => Quantity * Price;
         
 
         string Print()
